Introduce proto file patcher
Provide a new implementation for the file-based protobuf patcher,
building on top of the previous in-memory only implementation.

We need to be able to generate Protobuf files that can be read and
processed by the Protobuf compiler, so pure in-memory implementation is
not sufficient.
