+++ Cargo.toml
@@ -4,4 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+rayon = { version = "1.11.0", default-features = false }
+tempfile = { version = "3.21.0", default-features = false }
 thiserror = { version = "2.0.16", default-features = false, features = ["std"] }
+walkdir = { version = "2.5.0", default-features = false }
